Add program info to the bottom of the help page
Since the help page comes from the project's readme, the program info
should be in the help page only. This info should not be in the readme
because if someone can directly see the readme, then they can
easily find the program info elsewhere. Also, this info is already in
the package.json and should not be duplicated.